requirements.txt
streamlit
langchain
huggingface_hub
langchain_community
langchain_core
langchain_groq 
python-dotenv